Building Graphs

Introduction

We use LightOSM.jl as a basis to handle all the downloading, saving and importing of OSM data. Therefore, we expose very functions very similar to the ones in LightOSM.jl.

Then, the resulting LightOSM.OSMGraph instance is parsed into a MetaDiGraph, which is preserves the topology of the original street network, while reducing the number of nodes as much as possible. The geometry and various other parameters are attached to the props of every edge and vertex.

Even though we only expose a few functions for ease of use, we show all functions here, since it is very important to get an idea of what exactly this code is doing to your graph, before you start using it in any scientific capacity.

API

ShadowGraphs.add_edge_with_data!

ShadowGraphs.geolinestring

ShadowGraphs.get_neighbor_osm_ids

ShadowGraphs.get_node_list

ShadowGraphs.get_rotational_direction

ShadowGraphs.is_circular_way

ShadowGraphs.is_end_node

ShadowGraphs.is_lolipop_node

ShadowGraphs.nodelist_between

ShadowGraphs.offset_point_between

ShadowGraphs.parse_lanes

ShadowGraphs.parse_raw_ways

ShadowGraphs.point_on_radius

ShadowGraphs.shadow_graph_from_download

ShadowGraphs.shadow_graph_from_file

ShadowGraphs.shadow_graph_from_light_osm_graph

ShadowGraphs.shadow_graph_from_object

ShadowGraphs.width

ShadowGraphs.add_edge_with_data! — Method

add_edge_with_data!(g, s, d; data=Dict())

adds new edge from s to d to g::MetaDiGraph, and populates it with the props given in data.

Special care is given to self and multi edges:

self edges: if s==d, actually two new vertices with props of :lat, :lon, pointgeom and :helper=true are added.

these new vertices (h1 and h2) are then connected to form a loop like: s --he1--> h1 --real_edge--> h2 --he2--> d, where he1 and he2 are helper edges with only one prop of :helper=true. real_edge is carrying all the props defined in data

multi edges: if Edge(s,d) ∈ Edges(g), we add one new helper vertex with props of :lat, :lon, pointgeom and :helper=true.

We connect to the graph like this: s --he--> h --real_edge--> d, where he is a helper edge with only one prop, :helper=true. real_edge carries all the props specified in data

This process is nessecary to preserve the street network topology, since MetaDiGraphs do not support multi edges (and therefore also no multi self edges).

ShadowGraphs.geolinestring — Method

geolinestring(nodes, node_id_list)

creates an ArchGDAL linestring from a dictionary mapping osm node ids to LightOSM.Node and a list of osm node ids, representing the nodes of the linestring in order.

ShadowGraphs.get_neighbor_osm_ids — Method

get_neighbor_osm_ids(way::Way, start_id_index, nodes_in_nav_graph)

gets the osm ids of directly connected nodes in the reduced (topological) graph along the way. The starting node from which the neighbour ids are to be calculated is assumed to be at index start_id_index in the array nodes_in_nav_graph.

arguments

way: Way along which the neighbours are situated.

startidindex: index of the start node in nodes_in_nav_graph

nodes_in_nav_graph: array with osm ids of the nodes which form the way, which are also topologically relevant.

returns

tuple with:

array of neighbouring osm ids

array of directions (either +1 or -1`) that had to be taken from the start index to get to these neighbours

(if you go "along" the Way or "against" it).

notes

While we check if all nodes in nodes_in_nav_graph are in the way, we do not check if the order is correct.

In addition, we expect the user to take care of the case where the start and end of the way are the same node and in the nav graph. You basically want to get rid of either the start or the end duplicate if they are in the nav graph. If you have a way:

ring1 = Way(1, [10,20,30,40,50,60,70,80, 10], Dict("oneway"=>false, "reverseway"=>false, "name"=>"ring1"))

with the nodes [10, 30, 60, 80] in your nav graph, you need to pass only this list and not something like 10, 30, 60, 80, 10].

# correct
get_neighbor_osm_ids(ring1, 1, [10, 30, 60, 80])  # ([80, 30], [-1, 1])
get_neighbor_osm_ids(ring1, 1, [30, 60, 80, 10])  # ([10, 60], [-1, 1])

# danger
get_neighbor_osm_ids(ring1, 1, [10, 30, 60, 80, 10])  # ([10, 30], [-1, 1])

BUT! If you have a way like

loli1 = Way(1, [10,20,30,40,50,60,70, 30], Dict("oneway"=>false, "reverseway"=>false, "name"=>"loli1"))

you want to preserve the order and the duplicates. For example [10, 30, 60, 30], or, depending on the topology, maybe also [10, 30, 30]

ShadowGraphs.get_node_list — Method

get_node_list(way, start_pos, dest_osm_id, direction)

returns a list of all osm node ids between the start_pos and the destination node given by dest_osm_id, with step direction of direction. If either end of the array is reached during the steps, periodic boundaries are used.

arguments

way: Way whose nodes should be used

start_pos start index in way.nodes (NOT the osm_id of the start node)

dest_osm_id osm id of the destination node

direction direction in which the node list should be steped through in order to find the destination osm id.

returns

array with osm ids, with start id and destination id at start and end.

ShadowGraphs.get_rotational_direction — Method

get_rotational_direction(light_osm_graph)

calculates the dominant rotational direction of circular ways in light_osm_graph. This can be used to infere the side of the road on which people in a street network drive.

returns -1 if the rotation is lefthanded, 1 else. (prints warning if no clear direction could be established).

ShadowGraphs.is_circular_way — Method

is_circular_way(way::Way)

checks if a LightOSM.Way way starts at the same node it ends.

ShadowGraphs.is_end_node — Method

is_end_node(g, index)

checks if node index in graph g represents the end of a street (that is, has as most one neighbour).

ShadowGraphs.is_lolipop_node — Method

is_lolipop_node(g, osm_id)

checks if the node with osm_id is a "lolipop node" in the LightOSM.OSMGraph graph. A "lolipop node" is a node which occurs at the start/end of a way, as well as somewhere in the middle. (For example, the node 1 is considered a "lolipop node"" in the following way: 1-2-3-4-5-1-6-7)

ShadowGraphs.nodelist_between — Method

nodelist_between(way, start_osm_id, dest_osm_id, direction)

builds the list of osm node ids in the way which are between the start and destination osm id (inclusively), by taking steps through the list of nodes in the direction of direction.

If the start osm id occurs twice in the way, the shorter list is returned.

ShadowGraphs.offset_point_between — Method

offset_point_between(g, s, d)

returns x and y coordinates of a point between nodes s and d in graph g, using the :lon and :lat properties of the nodes. the point is (approximately) centered between s and p, and offset from the centerline by a random distance. (Use this function only to generate visually offset point, whose location is not relevant for anything except plotting).

ShadowGraphs.parse_lanes — Method

parse_lanes(tags::AbstractDict, tagname)

parses the value of the key tagname in tags, assuming it to be a numerical value describing a certain number of lanes. Returns the parsed number of lanes if the tag exists or missing if not.

ShadowGraphs.parse_raw_ways — Method

parse_raw_ways(raw_ways, network_type)

parses a list of dicts describing OSM Ways into LightOSM.Way instances. This function is a slightly modified version of the one used in LightOSM (parse_osm_network_dict), to be able to use our own, non-dafault value assuming parsers for the labels.

returns

a dictionary mapping osm_id to LightOSM.Way

ShadowGraphs.point_on_radius — Method

point_on_radius(x, y, r)

returns x and y coordinates of two random points on a circle with radius r around a point given by x and y. The two points have a fixed angluar offset of π/3 around the center. (This function assumes kartesian coordinates and euclidean distances. It is used to gernerate locations of helper nodes).

ShadowGraphs.shadow_graph_from_download — Method

function shadow_graph_from_download(download_method::Symbol;
                                    network_type::Symbol=:drive,
                                    metadata::Bool=false,
                                    download_format::Symbol=:json,
                                    save_to_file_location::Union{String,Nothing}=nothing,
                                    download_kwargs...)

downloads and builds the shadow graph from OSM.

arguments

download_method::Symbol: Download method, choose from :place_name, :bbox or :point.

network_type::Symbol=:drive: Network type filter, pick from :drive, :drive_service, :walk, :bike, :all, :all_private, :none, :rail

metadata::Bool=false: Set true to return metadata.

download_format::Symbol=:json: Download format, either :osm, :xml or json.

save_to_file_location::Union{String,Nothing}=nothing: Specify a file location to save downloaded data to disk.

Required Kwargs for each Download Method

download_method=:place_name

place_name::String: Any place name string used as a search argument to the Nominatim API.

download_method=:bbox

minlat::AbstractFloat: Bottom left bounding box latitude coordinate.

minlon::AbstractFloat: Bottom left bounding box longitude coordinate.

maxlat::AbstractFloat: Top right bounding box latitude coordinate.

maxlon::AbstractFloat: Top right bounding box longitude coordinate.

download_method=:point

point::GeoLocation: Centroid point to draw the bounding box around.

radius::Number: Distance (km) from centroid point to each bounding box corner.

download_method=:polygon

polygon::AbstractVector: Vector of longitude-latitude pairs.

Network Types

:drive: Motorways excluding private and service ways.

:drive_service: Motorways including private and service ways.

:walk: Walkways only.

:bike: Cycleways only.

:all: All motorways, walkways and cycleways excluding private ways.

:all_private: All motorways, walkways and cycleways including private ways.

:none: No network filters.

:rail: Railways excluding proposed and platform.

returns

MetaDiGraph with topologically relevant nodes and edges and relevant data attached to every node and edge.

ShadowGraphs.shadow_graph_from_file — Method

shadow_graph_from_file(file_path::String; network_type::Symbol=:drive)

builds the shadow graph from a file containing OSM data. The file could have been downloaded with either shadow_graph_from_download or download_osm_network.

arguments

file_path: path to file. either .osm, .xml or .json

network_type: type of network stored in file. Options are the same as in LightOSM:

:drive, :drive_service, :walk, :bike, :all, :all_private, :none, :rail

ShadowGraphs.shadow_graph_from_light_osm_graph — Method

shadow_graph_from_light_osm_graph(g)

transforms a LightOSM.OSMGraph into a MetaDiGraph, containing only the topologically relevant nodes and edges. Attached to every edge and node comes a lot of data, describing this specific edge or node:

nodes

in the case of helper nodes:

:lat

:lon

pointgeom

:helper=true

in the case of non helper nodes:

:osm_id

:lat

:lon

pointgeom

:helper=false

edges

in the case of helper edges:

:helper=true

in the case of non helper edges:

:osm_id

:tags (tags of the original osm way, with parsed width, lanes, lanes:forward, lanes:backward and lanes:both_ways, oneway and reverseway keys)

:edgegeom (ArchGDAL linestring with the geometry of the edge)

:geomlength=0

:parsing_direction (direction in which we stepped through the original way nodes to get the linestring)

:helper=false

ShadowGraphs.shadow_graph_from_object — Method

shadow_graph_from_object(osm_data_object::Union{XMLDocument,Dict}; network_type::Symbol=:drive)

builds the shadow graph from an object holding the raw OSM data. This function is using the graph_from_object function from LightOSM to first build a LightOSM.OSMGraph object which then gets augmented with the custom parsed ways, before it gets handed over to the shadow_graph_from_light_osm_graph function.

arguments

osmdataobject

networktype: type of network stored in osmdata_object. Options are the same as in LightOSM:

:drive, :drive_service, :walk, :bike, :all, :all_private, :none, :rail

ShadowGraphs.width — Method

width(tags)

less opinionated version of the basic parsing LightOSM does, to parse the width tag of an osm way. Returns the parsed width if the tag exists or missing if not.