Graph Measures

Introduction

We reimplement some measures for spatial graphs presented in OSMnx to work on the shadow graphs described in this package.

For now, only bearing and orientation entropy are available. See Boeing, G. 2019. “Urban Spatial Order: Street Network Orientation, Configuration, and Entropy.” Applied Network Science, 4 (1), 67 for the technical details.

Whenever a measurment attaches data to the graph, the tag will be prefixed with ms_, generally we do not checked if the tag exists before overwriting it. Modifying the graph will cause the measures depending on the modified part to go out of sync.

API

ShadowGraphs.bearing_histogram

ShadowGraphs.orientation_entropy

ShadowGraphs.orientation_order

ShadowGraphs.single_bearing

ShadowGraphs.tag_edge_bearings!

ShadowGraphs.bearing_histogram — Method

bearing_histogram(g::AbstractMetaGraph; weight=nothing, nbins=36, binshift=-180 / nbins, refresh_bearings=false)

fits a histogram to the bearings stored in the :ms_bearing properties of edges. If refresh_bearings=true, the bearings are updated/initialised before fitting.

keyword arguments

weight: name of property on the graphs edges to use as weights for the histogram. (for examle :sg_street_length. Is set to nothing, weight will be 1 for each edge.)

nbins: number of bins for the histogram.

binshift: offset for bins, in degrees. (use this if your bearings end up on bin edges.)

refresh_bearings: if the bearings should be (re-) calculated before fitting the histogram.

The resulting bins will span from binshift to 360+binshift.

returns count histogram of bearings.

ShadowGraphs.orientation_entropy — Method

orientation_entropy(hist::Histogram)

calculates the entropy of the (continuous) distribution approximated by hist according to:

$H=-\int{\rho\ln(\rho)\mathrm{d}x} \approx -\sum{\rho_i\ln(\rho_i)\Delta x_i}$

where $\Delta x_i$ is the width of the i-th bin.

There is nothing inherently orientation about this function. It should just work with any general distribution you throw at it.

ShadowGraphs.orientation_order — Method

orientation_order(hist::Histogram)

calculates the orientation order of a hist, defined as:

$\varphi = 1-\left(\frac{H-H_{min}}{H_{max}-H_{min}}\right)^2$

where $H_{min}$ is the entropy of a perfect grid and $H_{max}$ is the entropy of the constant orientation distribution.

The grid entropy is only (somewhat) defined for distributions with constant bin widths, so we check if hist fulfills this condition. I am not sure if I would trust this value.

ShadowGraphs.single_bearing — Method

single_bearing(line)

calculates the bearing of a single ArchGDAL linestring. Assumes the line is in a local coordinate system.

ShadowGraphs.tag_edge_bearings! — Method

tag_edge_bearings!(g::AbstractMetaGraph)

calculates the bearing of each :sg_street_geometry and attaches it to the respective edge as :ms_bearing. Helpers and self edges are ignored.