Pedestrianization

Introduction

Our shadow graphs are generally directed to allow for one-way streets. However, in the context of pedestrian routing, we might want to ignore this directedness. When pedestrianizing a graph, we add the reverse edge for every edge if it does not yet exist, together with the correctly inverted data like geometry and parsing direction.

Apply before centerline correction, as the revese edge should be offset in the other direction.

API

ShadowGraphs.pedestrianize!

ShadowGraphs.reverse_geometry

ShadowGraphs.pedestrianize! — Method

pedestrianize!(g::MetaDiGraph)

for every edge in g which has not a reverse edge, adds a reverse edge with reversed geometry (:sg_geometry_base and sg_street_geometry) and parsing direction :sg_parsing_direction * -1 of the original one. For :sg_helper edges, adds only a key :sg_helper=true to the new reverse edge.

Apply before centerline correction, as the new reverse geometry will be corrected in the other direction.

ShadowGraphs.reverse_geometry — Method

reverse_geometry(geom)

returns a copy of the input linestring with the same spatialref but reversed order of points.