Persistence

Introduction

Here we collect functions responsible for saving and loading ShadowGraphs to multiple CSV files with fine control over which properties get saved. This is useful if you want to persist some state of your graph, for example after adding the shadow intervals, but you are only interested in sg_street_length and sg_shadow_length.

There is a very rudimentary loader returning DataFrames, which might be a good starting point to build a graph from a persisted state as outlined above.

API

ShadowGraphs.export_shadow_graph_to_csv

ShadowGraphs.import_shadow_graph_from_csv

ShadowGraphs.export_shadow_graph_to_csv — Method

export_shadow_graph_to_csv(path, graph; edge_props=All(), vertex_props=All(), graph_props=All())

saves the shadow graph to a selection of csv files:

"path"_nodes.csv

"path"_edges.csv

"path"_graph.csv (contains graph properties)

All ArchGDAL geometries are converted to WellKnownText.

arguments

path: path to the target directory. The different specifiers are appended to the filename given in this path.

graph::MetaDiGraph: shadow graph to save.

keyword arguments

edge_props

vertex_props

graph_props

are used to select which of the props on each edge, vertex and graph should be exported. Each argument takes either:

DataFrames.All(): stores every property present.

DataFrames.Not([...]): stores every porperty except for the property names passed as a vector to Not. (example: Not([:sg_helper, :sg_geometry_base]))

[...]: only exports the properties with names in the vector. (example: [:sg_lon, :sg_lat, :sg_helper])

If an element does not have the property, an empty cell gets inserted in the csv, but only if there is at least one element with the property.

returns

saves multiple files to disk.

ShadowGraphs.import_shadow_graph_from_csv — Method

import_shadow_graph_from_csv(path)

imports csv files saved with export_shadow_graph_to_csv. path points to the main name of the files (without suffixes). (Just plug in the same thing you plugged in to save the graph).

returns

edge_df, vertex_df, graph_df: the loaded csv files as DataFrames to be further processed into whatever you need them to be.