Projecting Shadow Graphs

Introduction

We add methods to project_local! and project_back! from CoolWalksUtils.jl which allow for simpler projection of shadow graphs. Whithin them, the ArchGDAL geometry attached to nodes and edges are projected into a transverse mercator projection, and back.

API

CoolWalksUtils.project_back!

CoolWalksUtils.project_local!

ShadowGraphs._execute_edge_projection!

ShadowGraphs._execute_node_projection!

CoolWalksUtils.project_back! — Method

project_back!(g::T) where {T<:AbstractMetaGraph}

projects all ArchGDAL geometry in the props of edges and nodes of g back to EPSG4236 and updates the :sg_crs property of g accordingly. Returns g.

CoolWalksUtils.project_local! — Method

project_local!(g::T, obs::ShadowObservatory=get_prop(g, :sg_observatory)) where {T<:AbstractMetaGraph}
project_local!(g::T, lon, lat) where {T<:AbstractMetaGraph}

projects all ArchGDAL geometry in the props of edges and nodes of g to transverse mercator, centered around the obs (or lon, lat) and updates the :sg_crs property of g accordingly. All geometry is assumed to be in the :sg_crs system. Returns the projected g.

ShadowGraphs._execute_edge_projection! — Method

_execute_edge_projection!(g, trans)

applies the ArchGDAL transformation trans to every ArchGDAL geometry in the props of the edges of g.

ShadowGraphs._execute_node_projection! — Method

_execute_node_projection!(g, trans)

applies the ArchGDAL transformation trans to every ArchGDAL geometry in the props of the nodes of g.