Building Graphs

Introduction

We use LightOSM.jl as a basis to handle all the downloading, saving and importing of OSM data. Therefore, we expose very functions very similar to the ones in LightOSM.jl.

Then, the resulting LightOSM.OSMGraph instance is parsed into a MetaDiGraph, which is preserves the topology of the original street network, while reducing the number of nodes as much as possible. The geometry and various other parameters are attached to the props of every edge and vertex.

Even though we only expose a few functions for ease of use, we show all functions here, since it is very important to get an idea of what exactly this code is doing to your graph, before you start using it in any scientific capacity.

API

ShadowGraphs.add_edge_with_data!

ShadowGraphs.add_this_node

ShadowGraphs.countall

ShadowGraphs.decompose_way_to_primitives

ShadowGraphs.geolinestring

ShadowGraphs.get_node_list

ShadowGraphs.get_rotational_direction

ShadowGraphs.is_circular_way

ShadowGraphs.is_end_node

ShadowGraphs.parse_lanes

ShadowGraphs.parse_raw_ways

ShadowGraphs.shadow_graph_from_download

ShadowGraphs.shadow_graph_from_file

ShadowGraphs.shadow_graph_from_light_osm_graph

ShadowGraphs.shadow_graph_from_object

ShadowGraphs.width

ShadowGraphs.add_edge_with_data! — Method

add_edge_with_data!(g, s, d; data=Dict())

adds new edge from s to d to g::MetaDiGraph, and populates it with the props given in data.

data is expected to have at least a key :edgegeom, containing the geometry of the street between s and d as an ArchGDAL linestring in the WSG84 system. (Use apply_wsg_84!)

Special care is given to self and multi edges:

self edges: if s==d, actually two new vertices are added at 10% and 60% along the :edgegeom, with props of :lat, :lon, pointgeom and :helper=true set acordingly.

These new vertices (h1 and h2) are then connected to form a loop like: s --he1--> h1 --real_edge--> h2 --he2--> d, where he1 and he2 are helper edges with only one prop of :helper=true. real_edge is carrying all the props defined in data

multi edges: if Edge(s,d) ∈ Edges(g), we add one new helper vertex at 50% along the :edgegeom with props of :lat, :lon, pointgeom and :helper=true.

We connect to the graph like this: s --he--> h --real_edge--> d, where he is a helper edge with only one prop, :helper=true. real_edge carries all the props specified in data

This process is nessecary to preserve the street network topology, since MetaDiGraphs do not support multi edges (and therefore also no multi self edges).

ShadowGraphs.add_this_node — Method

add_this_node(g, osm_id)

checks if the node with osm_id in graph g should be added to the shadow graph. Churrently, we add a node if one of the following is true:

if the number of ways the node is part of is larger than 1

if the node is the end of a street, that is, if he has only one neighbour in g

if the node occurs more than once in the way it is part of, excluding the end point, if the way is circular

ShadowGraphs.countall — Method

countall(numbers)

counts how often every number appears in numbers. Returns dict with number=>count

ShadowGraphs.decompose_way_to_primitives — Method

decompose_way_to_primitives(way::Way)

Decomposed a way with possible self-intersections/loops into multiple ways which are guaranteed to be either

non-intersecting lines, where every node in the way is unique, or

circular ways, where only the first and last node in the way are not unique.

example

a simple way with nodes [10,20,30,40,50,30] (which looks like a triangle on a stick, not the repeated 30) will be decomposed into two ways, with the nodes [10,20,30] and [30,40,50,30].

ShadowGraphs.geolinestring — Method

geolinestring(nodes, node_id_list)

creates an ArchGDAL linestring from a dictionary mapping osm node ids to LightOSM.Node and a list of osm node ids, representing the nodes of the linestring in order.

ShadowGraphs.get_node_list — Method

get_node_list(simple_way, start_osm_id, topological_nodes, direction)

tries to get list of osm ids in simple_way between the startosmid and the next topologically relevant node in direction of direction. Returns either Int[] or nothing if there is no relevant node in direction of direction.

Used only on simplified ways.

ShadowGraphs.get_rotational_direction — Method

get_rotational_direction(way::Way, nodes, direction)

calculates the direction of rotation of the way if walked through in the direction of direction (either 1 or -1). Since the node locations are stored seperately, you have to pass in a dict node_id=>LightOSM.Node seperately. Used only on simplified ways.

Returns

1 if the rotation is righthanded

-1 if the rotation is lefthanded

0 if the way is not closed

ShadowGraphs.is_circular_way — Method

is_circular_way(way::Way)

checks if a LightOSM.Way way starts at the same node it ends.

ShadowGraphs.is_end_node — Method

is_end_node(g, index)

checks if node index in graph g represents the end of a street (that is, has as most one neighbour).

ShadowGraphs.parse_lanes — Method

parse_lanes(tags::AbstractDict, tagname)

parses the value of the key tagname in tags, assuming it to be a numerical value describing a certain number of lanes. Returns the parsed number of lanes if the tag exists or missing if not.

ShadowGraphs.parse_raw_ways — Method

parse_raw_ways(raw_ways, network_type)

parses a list of dicts describing OSM Ways into LightOSM.Way instances. This function is a slightly modified version of the one used in LightOSM (parse_osm_network_dict), to be able to use our own, non-dafault value assuming parsers for the labels.

returns

a dictionary mapping osm_id to LightOSM.Way

ShadowGraphs.shadow_graph_from_download — Method

function shadow_graph_from_download(download_method::Symbol;
                                    network_type::Symbol=:drive,
                                    metadata::Bool=false,
                                    download_format::Symbol=:json,
                                    save_to_file_location::Union{String,Nothing}=nothing,
                                    download_kwargs...)

downloads and builds the shadow graph from OSM.

arguments

download_method::Symbol: Download method, choose from :place_name, :bbox or :point.

network_type::Symbol=:drive: Network type filter, pick from :drive, :drive_service, :walk, :bike, :all, :all_private, :none, :rail

metadata::Bool=false: Set true to return metadata.

download_format::Symbol=:json: Download format, either :osm, :xml or json.

save_to_file_location::Union{String,Nothing}=nothing: Specify a file location to save downloaded data to disk.

Required Kwargs for each Download Method

download_method=:place_name

place_name::String: Any place name string used as a search argument to the Nominatim API.

download_method=:bbox

minlat::AbstractFloat: Bottom left bounding box latitude coordinate.

minlon::AbstractFloat: Bottom left bounding box longitude coordinate.

maxlat::AbstractFloat: Top right bounding box latitude coordinate.

maxlon::AbstractFloat: Top right bounding box longitude coordinate.

download_method=:point

point::GeoLocation: Centroid point to draw the bounding box around.

radius::Number: Distance (km) from centroid point to each bounding box corner.

download_method=:polygon

polygon::AbstractVector: Vector of longitude-latitude pairs.

Network Types

:drive: Motorways excluding private and service ways.

:drive_service: Motorways including private and service ways.

:walk: Walkways only.

:bike: Cycleways only.

:all: All motorways, walkways and cycleways excluding private ways.

:all_private: All motorways, walkways and cycleways including private ways.

:none: No network filters.

:rail: Railways excluding proposed and platform.

returns

MetaDiGraph with topologically relevant nodes and edges and relevant data attached to every node and edge.

ShadowGraphs.shadow_graph_from_file — Method

shadow_graph_from_file(file_path::String; network_type::Symbol=:drive)

builds the shadow graph from a file containing OSM data. The file could have been downloaded with either shadow_graph_from_download or download_osm_network.

arguments

file_path: path to file. either .osm, .xml or .json

network_type: type of network stored in file. Options are the same as in LightOSM:

:drive, :drive_service, :walk, :bike, :all, :all_private, :none, :rail

ShadowGraphs.shadow_graph_from_light_osm_graph — Method

shadow_graph_from_light_osm_graph(g)

transforms a LightOSM.OSMGraph into a MetaDiGraph, containing only the topologically relevant nodes and edges. Attached to every edge and node comes a lot of data, describing this specific edge or node:

nodes

in the case of helper nodes:

:lat

:lon

pointgeom

:helper=true

in the case of non helper nodes:

:osm_id

:lat

:lon

pointgeom

:helper=false

edges

in the case of helper edges:

:helper=true

in the case of non helper edges:

:osm_id

:tags (tags of the original osm way, with parsed width, lanes, lanes:forward, lanes:backward and lanes:both_ways, oneway and reverseway keys)

:edgegeom_base (ArchGDAL linestring with the geometry of the edge. Should not be modified during subsequent operations.)

:edgegeom (copy of :edgegeom_base. This one will be modified during offsetting and such.)

:full_length (length of edgegeom in a projected crs)

:parsing_direction (direction in which we stepped through the original way nodes to get the linestring)

:helper=false

the props '[:osmid, :tags, :edgegeombase, :parsingdirection, :helper] are considered read-only. (Editing them might cause strange behaviour. Always duplicate the `:edgegeombasewithArchGDAL.clone`)

ShadowGraphs.shadow_graph_from_object — Method

shadow_graph_from_object(osm_data_object::Union{XMLDocument,Dict}; network_type::Symbol=:drive)

builds the shadow graph from an object holding the raw OSM data. This function is using the graph_from_object function from LightOSM to first build a LightOSM.OSMGraph object which then gets augmented with the custom parsed ways, before it gets handed over to the shadow_graph_from_light_osm_graph function.

arguments

osmdataobject

networktype: type of network stored in osmdata_object. Options are the same as in LightOSM:

:drive, :drive_service, :walk, :bike, :all, :all_private, :none, :rail

ShadowGraphs.width — Method

width(tags)

less opinionated version of the basic parsing LightOSM does, to parse the width tag of an osm way. Returns the parsed width if the tag exists or missing if not.If Values are negative, we take the absolute value.