Persistence

Introduction

Here we collect functions responsible for saving and loading ShadowGraphs to multiple CSV files. Currently, only saving with two different levels of verbosity is supported.

API

ShadowGraphs.export_graph_to_csv

ShadowGraphs.import_graph_from_csv

ShadowGraphs.export_graph_to_csv — Method

export_graph_to_csv(path, graph; remove_internal_data = false)

saves the shadow graph to a selection of csv files:

"path"_nodes.csv

"path"_edges.csv

"path"_graph.csv (contains graph properties)

arguments

path: path to the target directory. The different specifiers are appended to the filename given in this path.

graph: shadow graph to save

remove_internal_data: whether to remove internal data used for future calculations. (set this to true, if you do not need to be able to reimport the graph into ShadowGraphs.jl and want just the relevant "exposed" data)

returns

saves multiple files to disk.

ShadowGraphs.import_graph_from_csv — Method

import_graph_from_csv(path)

imports graph saved via exportgraphto_csv. path points to the main name of the files (without suffixes). (Just plug in the same thing you plugged in to save the graph).