Plotting

Introduction

We extend the functionality of (Folium.jl)[https://github.com/SuperGrobi/Folium.jl] to be quickly able to visualise the shadowgraph in an interactive Leaflet.js map.

As most alternative packages in julia are quiet slow with regards to the drawing of very large graphs, this is currently the only way of getting close to interactive maps.

TODO:

maybe vectorise the inputs to set values "per edge"/ "per node" (GraphRecipes.jl?)

use/find alternative options to plot static (publication) level networks

API

Folium.draw!

Folium.draw!

Folium.draw! — Method

draw!(fig::FoliumMap, g::T, path::AbstractArray; kwargs...) where {T<:AbstractMetaGraph}

draws the path given by node ids in g into fig. Uses the :pointgeom-field of the nodes and the :edgegeom field of the edges. kwargs are applied to both, cirles and lines.

Folium.draw! — Method

draw!(fig::FoliumMap, g::T, series_type::Symbol; kwargs...) where {T<:AbstractMetaGraph}

draws the given data series properies of the graph into a Folium.jl map.

arguments

fig::FoliumMap: map to draw in

g: shadow graph carying the data

series_type: type of data to draw from the graph. Pick from: :vertices, :edges, :edgegeom, :shadowgeom.

drawarrows: if there should be a circle drawn 80% along the edgegeom. (only applies when `seriestypeis:edgegeom`)

kwargs: keywords passed to folium for every series_type. (see the python docs and the leaflet docs for a full list of all options.)

Every series_type has a few sensible defaults set, most importantly the default tooltips and popups, which, by default show some interesting data about the vertices and edges, respectively. Currently, the kwargs are set for every element, there is (currently) no way to set parameters with, for example a vector, to get a element by element arguments.

returns

fig::FoliumMap (passthrough of argument)