Projecting Shadow Graphs

Introduction

We extend the functionality of (CoolWalksUtils.jl)[https://github.com/SuperGrobi/CoolWalksUtils.jl] with functions to project all the ArchGDAL geometry attached to nodes and edges into a transverse mercator projection, to do geometric operations on them.

API

CoolWalksUtils.project_back!

CoolWalksUtils.project_local!

ShadowGraphs.project_graph_edges!

ShadowGraphs.project_graph_nodes!

CoolWalksUtils.project_back! — Method

project_back!(g::T) where {T<:AbstractMetaGraph}

projects all ArchGDAL geometry in the props of edges and nodes of g back to EPSG4236 and updates the :crs property of g accordingly.

CoolWalksUtils.project_local! — Method

project_local!(g::T, center_lon, center_lat) where {T<:AbstractMetaGraph}

projects all ArchGDAL geometry in the props of edges and nodes of g to transverse mercator, centered at center_lon, center_lat and updates the :crs properity of g accordingly. All geometry is assumed to be in the :crs system.

ShadowGraphs.project_graph_edges! — Method

project_graph_edges!(g, trans)

applies the ArchGDAL transformation trans to every ArchGDAL geometry in the props of the edges of g.

ShadowGraphs.project_graph_nodes! — Method

project_graph_nodes!(g, trans)

applies the ArchGDAL transformation trans to every ArchGDAL geometry in the props of the nodes of g.